refactored blockchain interface and restructured dependencies in the jsonrpc package
