+++ jsonrpc/jsonrpc.go
@@ -36,16 +36,25 @@ func (s serverType) String() string {
 type JSONRPC struct {
 	logger     hclog.Logger
 	config     *Config
-	dispatcher dispatcherImpl
+	dispatcher dispatcher
 }
 
-type dispatcherImpl interface {
+type dispatcher interface {
 	HandleWs(reqBody []byte, conn wsConn) ([]byte, error)
 	Handle(reqBody []byte) ([]byte, error)
 }
 
+// JsonRPCStore defines all the methods required
+// by all the JSON RPC endpoints
+type JsonRPCStore interface {
+	ethStore
+	networkStore
+	txPoolStore
+	filterManagerStore
+}
+
 type Config struct {
-	Store   blockchainInterface
+	Store   JsonRPCStore
 	Addr    *net.TCPAddr
 	ChainID uint64
 }
