+++ jsonrpc/net_endpoint.go
@@ -2,14 +2,20 @@ package jsonrpc
 
 import "strconv"
 
+// networkStore provides methods needed for Net endpoint
+type networkStore interface {
+	GetPeers() int
+}
+
 // Net is the net jsonrpc endpoint
 type Net struct {
-	d *Dispatcher
+	store   networkStore
+	chainId uint64
 }
 
 // Version returns the current network id
 func (n *Net) Version() (interface{}, error) {
-	return strconv.FormatUint(n.d.chainID, 10), nil
+	return strconv.FormatUint(n.chainId, 10), nil
 }
 
 // Listening returns true if client is actively listening for network connections
@@ -19,7 +25,7 @@ func (n *Net) Listening() (interface{}, error) {
 
 // PeerCount returns number of peers currently connected to the client
 func (n *Net) PeerCount() (interface{}, error) {
-	peers := n.d.store.GetPeers()
+	peers := n.store.GetPeers()
 
 	return strconv.FormatInt(int64(peers), 10), nil
 }
